changed a bunch of things to get absolute drive to work

+++ src/com/_2491nomythic/helios/settings/Constants.java
@@ -23,7 +23,7 @@ public class Constants {
 	public static final boolean driveEncoderCenterReversed = false;
 	public static final boolean driveEncoderRightReversed = false;
 	public static final double driveEncoderToFeet = 1.0;
-	public static final double gyroChannel = 1;
+	public static final int gyroChannel = 1;
 	
 	// Arm
 	public static final int armTalonArmMotorLeftChannel = 13;
